User gmdate() instead of date_i18n().
This is only for the legacy dates fields. Using date_i18n() was applying
the current timezone offset which was causing the dates to output
incorrectly because the dates were created with gmmktime() without a
timezone offset applied. From Google searches, if you use gmmktime() to
create a timestamp you should use gmdate() display it.