Fix bug in the template file override code.
Iterate thru each file name in every folder rather than iterating over every folder then every file. This is important because a lower priority file in a higher priority folder could override a higher priority file because it was in a lower priority folder. This should now check for the higher priority file in each folder before moving on to the next priority file.