Refactor cnTemplate::locate() to fix inconsistencies in the priority loading order of template override resource files.
The search should look for each file in priority order within the current folder before moving on to the next lower priority folder.

As part of this, the minified resources for CSS|JS files will be checked for in the current folder and loaded instead of the non-minified resource if it exists. The filter that added the minified resources to the filenames has been disabled and should be removed at a later date.